backend/transport_system/requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
drf-extra-fields==3.1.1
Django==4.0.4
django-colorfield==0.6.3
django-filter==2.4.0
djangorestframework==3.13.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drfpasswordless==1.5.7
flake8==4.0.1
flake8-isort==4.1.1
gunicorn==20.0.4
haversine==2.5.1
idna==2.10
importlib-metadata==1.6.0
isort==5.10.1
mccabe==0.6.1
more-itertools==8.2.0
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.9.3
py==1.10.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.0.1
pyparsing==2.4.7
pytest==7.1.1
pytest-django==4.5.0
pytz==2021.1
requests==2.25.1
six==1.14.0
sqlparse==0.4.2
svgwrite==1.4.2
toml==0.10.2
urllib3==1.26.5
wcwidth==0.1.9
wheel==0.36.2
zipp==3.1.0
